+++ pyproject.toml
@@ -13,6 +13,8 @@ dependencies = [
     "python-dotenv~=1.0",
     "kaggle~=1.5",
     "dbt-postgres>=1.4.5",
+    "pyarrow>=11.0.0",
+    "pydantic>=1.10.7",
     "apache-beam>=2.47.0",
     "apache-flink>=1.9.3",
 ]
@@ -44,9 +46,15 @@ testpaths = [
     "tests"
 ]
 python_functions = "should__*"
-python_files = "*_test.py"
+python_files = "test_*.py"
 
 [tool.pdm]
 
 [tool.dagster]
 module_name = "dagster_covid19.asset"
+
+[tool.ufmt]
+excludes = [
+    "covid19",
+    "dagster-home"
+]
